fix: Cria o arquivo app.ts e cópia o código de configuração do servidor que estava no arquivo index.ts.
Realizada a instalação do mysql2 para resolver o problema de autenticação. Mas, pelo que parece foi o fato da configuração do servidor estar conflitando com o código no arquivo index.ts. Agora é possível bater nos endpoints e gerar as requisições
