+++ modulo6/heranca-encapsulamento/package.json
@@ -1,7 +1,6 @@
 {
   "name": "template-revisao-modulo4",
   "version": "1.0.0",
-  "description": "",
   "main": "index.js",
   "scripts": {
     "start": "node ./build/index.js",
@@ -21,10 +20,11 @@
     "typescript": "^4.7.4"
   },
   "dependencies": {
+    "@mysql/xdevapi": "^8.0.32",
     "cors": "^2.8.5",
     "dotenv": "^16.0.1",
     "express": "^4.18.1",
     "knex": "^2.2.0",
-    "mysql": "^2.18.1"
+    "mysql2": "^3.2.0"
   }
 }
