+++ modulo6/heranca-encapsulamento/src/app.ts
@@ -1,0 +1,15 @@
+import express from 'express'
+import cors from 'cors'
+import dotenv from 'dotenv'
+
+dotenv.config()
+const app = express()
+
+app.use(express.json())
+app.use(cors())
+
+app.listen(process.env.PORT || 3003, () => {
+    console.log(`Servidor rodando na porta ${process.env.PORT || 3003}`)
+})
+
+export default app;
