+++ modulo6/heranca-encapsulamento/src/database/connection.ts
@@ -4,7 +4,7 @@ import dotenv from "dotenv"
 dotenv.config()
 
 const connection = knex({
-   client: "mysql",
+   client: "mysql2",
    connection: {
       host: process.env.DB_HOST,
       port: 3306,
@@ -15,4 +15,4 @@ const connection = knex({
    },
 });
 
-export default connection
+export default connection;
